Adjust project communication templates
We standardized the languages used in the communication of the project,
with Portuguese being used as the outer-facing language, for pull
requests and new issues, and English in the code and commit messages.
Due to that, we needed to update the templates we previously had.

Along with the translation, we decided to add a pull request template to
the project, since we had people write the PRs in their own way. This
helps with standardization and also speeds up the writing process.